error fix for migration

+++ core/Modules/Attributes/Database/Migrations/2022_07_25_090052_add_column_to_sub_categories_table.php
@@ -27,7 +27,14 @@ return new class extends Migration
     public function down()
     {
         Schema::table('sub_categories', function (Blueprint $table) {
-            $table->dropColumn("status_id");
+            try {
+                $table->dropForeign(['status_id']);
+            } catch (\Exception $e) {
+            }
+
+            if (Schema::hasColumn('sub_categories', 'status_id')) {
+                $table->dropColumn('status_id');
+            }
         });
     }
 };
