+++ core/Modules/Order/Database/Migrations/2022_12_14_131343_add_user_id_to_orders_table.php
@@ -16,7 +16,14 @@ return new class extends Migration {
     public function down()
     {
         Schema::table('orders', function (Blueprint $table) {
-            $table->dropColumn("user_id");
+            try {
+                $table->dropForeign(['user_id']);
+            } catch (\Exception $e) {
+            }
+
+            if (Schema::hasColumn('orders', 'user_id')) {
+                $table->dropColumn('user_id');
+            }
         });
     }
 };
