+++ core/database/migrations/2021_10_23_111023_add_sold_count_in_products_table.php
@@ -25,8 +25,12 @@ class AddSoldCountInProductsTable extends Migration
      */
     public function down()
     {
-        Schema::table('products', function (Blueprint $table) {
-            $table->dropColumn('sold_count');
-        });
+        if (Schema::hasTable('products')) {
+            Schema::table('products', function (Blueprint $table) {
+                if (Schema::hasColumn('products', 'sold_count')) {
+                    $table->dropColumn('sold_count');
+                }
+            });
+        }
     }
 }
